[qofonoext] Switch Nci property type as string. JB#61312
The qint64 type was problematic:
- Doesn't work in QML
- The Invalid64 on enum doesn't work without declaring the
enum as 64 bit. Which again might not be the best idea for qml.
- App developer needs to pay close attention whether to use
32-bit or 64-bit invalid value.

As this isn't getting used in the wild yet, let's just switch
the nci property as string for now. If there are c++ needs, a
separate getter can be added later.

The general 64-bit storage just because of one wider property, now
needing more special handling, felt heavy. Switched back to 32-bit and
using separate property for nci. Wouldn't expect new radio
technologies with new properties too soon.